tpm/requirements.txt
requests>=2.28.0
azure-identity>=1.15.0
